05_training/requirements.txt
# Core dependencies for ReFocused-AI 1B model training

# Deep Learning
torch>=2.1.0
transformers>=4.35.0
deepspeed>=0.12.0
accelerate>=0.25.0

# Google Cloud
google-cloud-storage>=2.10.0
google-auth>=2.23.0

# Data processing
numpy>=1.24.0
pandas>=2.1.0

# Monitoring and logging
wandb>=0.16.0
tensorboard>=2.15.0
gpustat>=1.1.1
tqdm>=4.66.0

# Optimization libraries
bitsandbytes>=0.41.0  # For 8-bit optimizers
einops>=0.7.0  # For efficient tensor operations

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
rich>=13.7.0  # For pretty console output
psutil>=5.9.0  # For system monitoring

# Development tools
black>=23.0.0
isort>=5.12.0
pytest>=7.4.0 